Implement Distance business object and its Dialog Picker

+++ app/src/main/java/au/com/beba/runninggoal/persistence/RunningGoalEntity.kt
@@ -8,29 +8,29 @@ import android.arch.persistence.room.PrimaryKey
 @Entity(tableName = "running_goal")
 data class RunningGoalEntity(
 
-    // REF https://developer.android.com/reference/android/arch/persistence/room/PrimaryKey#autoGenerate()
-    @PrimaryKey(autoGenerate = true)
-    var uid: Int = 0,   // IF SET AS 0 AUTO-INCREMENT WILL ASSIGN THE VALUE
+        // REF https://developer.android.com/reference/android/arch/persistence/room/PrimaryKey#autoGenerate()
+        @PrimaryKey(autoGenerate = true)
+        var uid: Int = 0,   // IF SET AS 0 AUTO-INCREMENT WILL ASSIGN THE VALUE
 
-    @ColumnInfo(name = "goal_name")
-    var goalName: String = "",
+        @ColumnInfo(name = "goal_name")
+        var goalName: String = "",
 
-    @ColumnInfo(name = "target_distance")
-    var targetDistance: Int = 0,
+        @ColumnInfo(name = "target_distance")
+        var targetDistance: Float = 0.0f,
 
-    @ColumnInfo(name = "current_distance")
-    var currentDistance: Double = 0.00,
+        @ColumnInfo(name = "current_distance")
+        var currentDistance: Float = 0.0f,
 
-    @ColumnInfo(name = "start_date")
-    var startDate: Long = 0,
+        @ColumnInfo(name = "start_date")
+        var startDate: Long = 0,
 
-    @ColumnInfo(name = "end_date")
-    var endDate: Long = 0,
+        @ColumnInfo(name = "end_date")
+        var endDate: Long = 0,
 
-    // TODO: CHANGE TO ONE-TO-MANY RELATIONSHIP
-    @ColumnInfo(name = "widget_id")
-    var widgetId: Int = 0,
+        // TODO: CHANGE TO ONE-TO-MANY RELATIONSHIP
+        @ColumnInfo(name = "widget_id")
+        var widgetId: Int = 0,
 
-    @ColumnInfo(name = "view_type")
-    var viewType: Int = 0
+        @ColumnInfo(name = "view_type")
+        var viewType: Int = 0
 )
